Update dndtable.sty so header row transparent
The table header row wasn't transparent, like it is in the DnD books. This looked off, and was bugging me, but nothing I did with \rowcolor seemed to fix it. I researched the problem and found this solution, by the author of tabularx no less: http://tex.stackexchange.com/questions/297345/why-is-the-start-row-of-rowcolors-ignored-in-tabularx