add quotebox color and clean up themed colors
Add quoteboxcolor. Removed trimcolor as I realized there really was not
any reason for a user to not use themecolor instead. Removed optional
color from paperbox and commentbox as the few times a user might want to
change the color of a single commentbox or paperbox can set the
tcolorbox relevant options.