Ensure all themed environments have the same color parameter.
Decided @benwebber was right. A unified means of changing the color of a
single instance of a themed environment is more important than cleaning
up the definition of an environment. It means that quotebox has the same
annoyance of \begin{quotebox}[][color] to change the color without any
other parameters being changed, but I do not see a good way around that.