dndfonts.sty: Fix compilation with xelatex and lualatex
dndfonts.sty includes kpfonts, which is compatible with pdflatex, but
not with xelatex and lualatex. While the kpfonts package defaults to
including kpfonts-otf when it detects xelatex or lualatex, it doesn't
pass options to kpfonts-otf, which makes lualatex and xelatex complain
about unrecognised options used in dndfonts.sty (i.e., notext, nomath,
nott), and change almost all fonts in the style.

This fix tests for pdflatex in lib/dndfonts.sty, and else loads
kpfonts-otf directly, correctly passing the necessary options.